third_party/WebKit/LayoutTests/editing/execCommand/insert_list_empty_div.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable>|</div>',
  'InsertUnorderedList',
  [
  '<div contenteditable>',
    '<ul><li>|<br></li></ul>',
  '</div>'
  ]),
  'This tests InsertList inside an empty div with no placeholder');
</script>
